// autotunex/ux/src/lib/app-types.ts
export type Job = {
	id: string;
	user_id: string;
	status: Status;
	seed: number;
	config_id: string;
	dataset_id: string;
	model: string;
	model_source: string;
	experiment_name: string;
	tuning_type: string;
	ray_address: null;
	precision: string;
	cleanup: number;
	autotune: number;
	created_at: Date;
	updated_at: Date;
};

export type Task = {
	id: string;
	job_id: string;
	build_id: string;
	status: string;
	type: string;
	pr_url: string;
	artifact_id: string;
	artifact_uri: string;
	build_status: BuildStatus;
	started_at: Date;
	updated_at: Date;
	rits_url: null;
};

export type Tuning = {
	user: string;
	config_name: string;
	dataset: string;
	num_trials: number;
	task_id: string;
	build_id: string;
	task_status: Status;
	task_type: string;
	github_pr_url: string;
	artifact_id: string;
	artifact_uri: string;
	task_started_at: Date;
	task_updated_at: Date;
	rits_url: null;
	trials?: Trial[];
	dmf?: any;
	assets?: Assets[];
	is_published?: boolean;
	build_status?: BuildStatus;
	logs?: Log[];
	gb_logs?: string[];
} & Job;
// ...
export type Pair = {
	input: any;
	output: any;
};
export type Dataset = {
	id: string;
	user_id: string;
	name: string;
	description: string;
	train_file: string;
	train_records: number;
	train_file_size: number;
	validation_file: string;
	validation_records: number;
	validation_file_size: number;
	artifact_id: string;
	artifact_url: string;
	created_at: Date;
	updated_at: Date;
	associated_jobs: Job[];
	validation_data?: Pair[];
	train_data?: Pair[];
};

export interface Configuration {
	id: string;
	user_id: string;
	name: string;
	tuner_type: string;
	rl_tuner_type?: string | null;
	artifact_id: string;
	artifact_url: string;
	config_data: ConfigData;
	associated_jobs: Job[];
	created_at: Date;
	updated_at: Date;
}

export type ConfigForm = {
	name?: string;
	tuner_type?: string;
	rl_tuner_type?: string;
} & ConfigData;

export interface ConfigData {
	tune_config: TuneConfig;
	tuners_config: TunersConfig;
	training_config: TrainingConfig;
	training_rl_config?: TrainingRlConfig;
	tuners_rl_config?: TunersRlConfig;
}

export interface TrainingConfig {
	[key: string]: HpoDatasetPercentage | InputColumn;
	seed: HpoDatasetPercentage;
	precision: InputColumn;
	max_length: HpoDatasetPercentage;
	input_column: InputColumn;
	warmup_ratio: HpoDatasetPercentage;
	output_column: InputColumn;
	hpo_num_epochs: HpoDatasetPercentage;
	num_train_epochs: HpoDatasetPercentage;
	use_chat_template: InputColumn;
	num_gpus_per_trial: HpoDatasetPercentage;
	num_cpus_per_worker: HpoDatasetPercentage;
	use_flash_attention: InputColumn;
	train_implementation: InputColumn;
	hpo_dataset_percentage: HpoDatasetPercentage;
}
// ...
export interface TuneConfig {
	[key: string]: InputColumn | HpoDatasetPercentage | NumberInputColumn | undefined;
	scheduler: InputColumn;
	search_alg: InputColumn;
	num_samples: HpoDatasetPercentage;
	max_discrepancy: HpoDatasetPercentage;
	max_concurrent_trials: HpoDatasetPercentage;
	time_budget_s?: NumberInputColumn;
}

export interface TunersConfig {
	[key: string]: Tuner;
	lora: Tuner;
	alora: Tuner;
}

export interface TunersRlConfig {
	[key: string]: Tuner;
}

export interface TrainingRlConfig {
	[key: string]: InputColumn | NumberInputColumn;
}
// ...
export type Trial = {
	id: string;
	job_id: string;
	status: Status;
	config: Record<string, any>;
	created_at: Date;
	updated_at: Date;
	logs: any[];
	score: Score;
};

export type Metric = 'loss';

export type Status = 'COMPLETED' | 'ERROR' | 'RUNNING' | 'TERMINATED' | 'PENDING' | 'SUBMITTED';

export type Score = {
	id: string;
	job_id: string;
	trial_id: string;
	metric: Metric;
	metrics: Metrics;
	created_at: Date;
	updated_at: Date;
};
// ...
export type Resources = {
	model_size_billion_params: number;
	gpu_memory_gb: number;
	cpu_memory_gb: number;
	num_gpus: number;
	weights_memory: number;
	optimizer_memory: number;
	gradients_memory: number;
	activations_memory: number;
};

export type Estimation = {
	model_name: string;
	config_id: string;
	gpu_memory: number;
};
// ...
export type Log = {
	id?: number;
	filename: string;
	level: 'INFO' | 'ERROR' | 'DEBUG' | 'WARN';
	message: string;
	timestamp: Date;
};

export type Assets = {
	path: string;
	size: number;
	created: Date;
	filename: string;
	modified: Date;
	file_hash: string;
	file_size: number;
	published: boolean;
};

export interface BuildStatus {
	details: Details;
	targets: Target[];
	build_history: BuildHistory[];
}

export interface BuildHistory {
	time: Date;
	description: string;
}

export interface Details {
	name: string;
	status: string;
	build_id: string;
	source_pr: string;
	started_at: Date;
	updated_at: Date;
}

export interface Target {
	steps: Step[];
	status: string;
	build_id: string;
	target_id: string;
	target_name: string;
	input_artifacts: PutArtifact[];
	output_artifacts: PutArtifact[];
}